Enforce the Brunch transport boundary in CI
Make transport dependency violations affect the core architecture task and fail during installation.
