+++ .yarnrc.yml
@@ -2,6 +2,8 @@ approvedGitRepositories: []
 
 defaultSemverRangePrefix: ""
 
+enableConstraintsChecks: true
+
 enableScripts: false
 
 enableTransparentWorkspaces: false
