+++ apps/petrinaut-website/package.json
@@ -19,6 +19,7 @@
   },
   "dependencies": {
     "@ai-sdk/openai": "3.0.63",
+    "@hashintel/brunch-agent": "workspace:*",
     "@hashintel/brunch-agent-transport-aisdk": "workspace:*",
     "@hashintel/ds-components": "workspace:*",
     "@hashintel/ds-helpers": "workspace:*",
