+++ apps/petrinaut-website/src/main/app/local-storage-demo/brunch-ask-interactive-tool.tsx
@@ -6,7 +6,7 @@ import {
   type BrunchAskOutput,
   parseBrunchAskInput,
   parseBrunchAskOutput,
-} from "@hashintel/brunch-agent-transport-aisdk/client-tools";
+} from "@hashintel/brunch-agent/client-tools";
 import { css } from "@hashintel/ds-helpers/css";
 import {
   definePetrinautAiInteractiveTool,
