+++ libs/@hashintel/brunch-agent/packages/transport-aisdk/vite.config.ts
@@ -8,9 +8,6 @@ export default defineConfig({
   build: {
     lib: {
       entry: {
-        "client-tools": fileURLToPath(
-          new URL("src/client-tools.ts", import.meta.url),
-        ),
         headers: fileURLToPath(new URL("src/headers.ts", import.meta.url)),
         index: fileURLToPath(new URL("src/index.ts", import.meta.url)),
       },
@@ -18,7 +15,7 @@ export default defineConfig({
       formats: ["es"],
     },
     rolldownOptions: {
-      external: [/^@hashintel\/brunch-agent(?:\/.*)?$/u, "ai", "valibot"],
+      external: ["ai", "valibot"],
     },
     sourcemap: true,
   },
